requirements.txt
yt-dlp>=2024.1.0
easyocr>=1.7.0
flask>=3.0.0
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
